+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/logs_collector_functions/package_private_helpers.go
@@ -19,6 +19,10 @@ import (
 
 const (
 	shouldShowStoppedLogsCollectorContainers = true
+
+	//This port number was used when whe run the logs collector without publishing the TCP port
+	//now we need to check for this, in order to know if the logs collector does not bind
+	oldLogsCollectorTCPPortNumber = 24224
 )
 
 func getLogsCollectorPrivatePorts(containerLabels map[string]string) (
@@ -93,13 +97,22 @@ func getLogsCollectorObjectFromContainerInfo(
 			return nil, stacktrace.NewError("Couldn't parse private IP address string '%v' to an IP", privateIpAddrStr)
 		}
 
-		publicIpAddr, publicTcpPortSpec, err = shared_helpers.GetPublicPortBindingFromPrivatePortSpec(privateTcpPortSpec, allHostMachinePortBindings)
+		// TODO Remove this after 2022-12-27, when we're confident nobody will have the old port
+		//checking for old port spec because old logs collectors container (older than engine v.0.50.3) do not publish the TCP port
+		isPrivateTCPPortFromOldLogsCollectorContainers, err := isLogsCollectorPortSpecFromOlbLogsCollectorContainers(privateTcpPortSpec, allHostMachinePortBindings)
 		if err != nil {
-			return nil,
-			stacktrace.Propagate(err, "The logs collector is running, but an error occurred getting the public port spec for the TCP private port spec '%+v'", privateTcpPortSpec)
+			return nil, stacktrace.Propagate(err, "An error occurred cheking if logs collector private TCP port spec '%+v' corresponds to and old version", privateTcpPortSpec)
 		}
 
-		_, publicHttpPortSpec, err = shared_helpers.GetPublicPortBindingFromPrivatePortSpec(privateHttpPortSpec, allHostMachinePortBindings)
+		if !isPrivateTCPPortFromOldLogsCollectorContainers {
+			_, publicTcpPortSpec, err = shared_helpers.GetPublicPortBindingFromPrivatePortSpec(privateTcpPortSpec, allHostMachinePortBindings)
+			if err != nil {
+				return nil,
+					stacktrace.Propagate(err, "The logs collector is running, but an error occurred getting the public port spec for the TCP private port spec '%+v'", privateTcpPortSpec)
+			}
+		}
+
+		publicIpAddr, publicHttpPortSpec, err = shared_helpers.GetPublicPortBindingFromPrivatePortSpec(privateHttpPortSpec, allHostMachinePortBindings)
 		if err != nil {
 			return nil, stacktrace.Propagate(err, "The logs collector is running, but an error occurred getting the public port spec for the HTTP private port spec '%+v'", privateHttpPortSpec)
 		}
@@ -119,6 +132,23 @@ func getLogsCollectorObjectFromContainerInfo(
 	)
 
 	return logsCollectorObj, nil
+}
+
+func isLogsCollectorPortSpecFromOlbLogsCollectorContainers(
+	privatePortSpec *port_spec.PortSpec,
+	allHostMachinePortBindings map[nat.Port]*nat.PortBinding,
+) (bool, error) {
+	// Convert port spec protocol -> Docker protocol string
+	dockerPort, err := shared_helpers.GetDockerPortFromPortSpec(privatePortSpec)
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred checking if this private TCP port spec '%+v' corresponds to an old version of the logs collector container", privatePortSpec)
+	}
+
+	_, found := allHostMachinePortBindings[dockerPort]
+	if !found && privatePortSpec.GetNumber() == oldLogsCollectorTCPPortNumber {
+		return true, nil
+	}
+	return false, nil
 }
 
 func getAllLogsCollectorContainers(ctx context.Context, dockerManager *docker_manager.DockerManager) ([]*types.Container, error) {